Speed things up; fix debug output on parse failure
There was a slowdown due to many recursive calls to repr after I added that
to Object; it was initiated by a temp debug print when running a user function,
so I just commented out the related lines.

For printing out parse failures, we really want to always track the farthest
point we've reached in a file. I tried clearing this data after each global
phrase success, but even that is a mistake since the parse error may really be
a continuation of a badly-ended-early phrase parse.