Move sample{11,12} to new mode syntax; speed fix.
I have to be careful with my dprint statements. Even when they're not
actually printed, they can easily be a bottleneck simply because
unlimited stringification of objects takes time. We're often dealing
with surprisingly large objects. In the future I may want to address
that, but maybe not until we switch to a custom language.