Add 13.water with easier whitespace ignoring.
Includes that string mode (for js_str) temporarily turns off whitespace
skipping. I found that one problem with this approach is that the prefix
is automatically affecting the lang def grammar itself. It would be good
to set up a mechanism to localize the prefix to a subset of the total
grammar being parsed.