Methods work on combined rules.
Also made _src -> src a public function and changed mode_result
to always be a list. As a follow-up, I think I have to adjust
some of the module code to make sure it still works with things
as they are.

Which keeps reminding me that I'd love to have unit tests for
all this stuff.