Add start of groups to layer2.water; not done.
It's tricky to add groups to combined rules. If combined rules were
built within the parser, we could call that parser function. But it
doesn't seem right to me that we'd have to invent such a complex
function, even in env, when it is relatively easy to handle within the
parser itself.

I'm not 100% of the best solution. For now, I like the idea of using a
loadstring-like function that can create a subparser and parse in
combined rules that we make up on the fly - sort of pausing the current
main parse stack to subparse a new top-level string. I want to have that
function eventually anyway, and it will work here. The trouble is that
it will probably be a chunk of work to get working well.