Add initial grouping grammar.
I worked around the subst issue without a loadstring by integrating
grouping in with the layer 2 grammar for combined rules. I decided that
this was a better approach than to put in a too-many-optioned loadstring
that could handle this particular case. I will eventually want
loadstring, but it can wait. And probably won't be called loadstring.