Start adding prefix_change grammar.
The current subgoal is to get 52.water working.
Next up is adding a prefix stack, and having _parse_item understand
prefix changes, and work with those.

I could also consider, long-term, a way to avoid this beign part
of _parse_item, but rather simply part of the api that is wrapped
in one of the non-core grammar layers.