Start setting up prefix stack in parse.py.
The old prefix system is not a stack; new one is.

I'll update parse.py when the new stuff is functional. I have
prase.pending.py as a separate temporary file because I don't want to
use a branch (I just choose not to for now), yet I want to keep the
primary code working decently. Maybe strange choice, but I like it.