Get a+b syntax working.
With some useful notes that temporarily live in layer2.water.

This commit also modifies 60.water to use the friendlier syntax.

Next up: a*b syntax, which is a short jump from here.

I'm also interested in soonly avoiding unnecessary rule expansions,
as in inner rules that expand a single item. (It's important that
they're inner rules due to the case of a single-item OrRule that someone
may later want to prepend to.)