Work toward fixed a+*b syntax.
The problem is that we weren't returning nice lists.

The current file doesn't quite work, but I think is the start of the
right approach. Next up I would like to consolidate the inner rules used
as the list elements. In general, I want to get 60.water working, and
then a+b syntax, and then think about factoring these together cleanly.