Work toward a@b syntax.
This is incomplete. The plan is to have a parsed a@b item turn into a
tuple which is recognized from _parse_item because the first tuple
element begins with '@'.

I originally was planning to simply separate the a@b item into a and b
again within parse.py, but I realized there was enough work in that to
handle strings (or other things I didn't even think of?) that it seems
cleaner to just parse it once, and keep a and b separate from that point
forward. In fact, it may even make sense to do this for labels as well,
although I'm not going to worry about that right now.