Officially add a@b syntax in layer1.
However, this breaks 62.water. Well, really 62.water was broken all
along and switching a+*b to use a@b internally will enable a fix for it.
Next up I think I'll make sure 63.water works, and then move to the fix
in layer2 that will get 62.water working as well.