Checkpoint while updating a+*b system.
This leaves at least 61.water in a broken state. I'd like to think
carefully about how lists work with prefixes. Probably the prefixes
should always stay on the elements, as opposed to introducing attributed
lists. The code (I think) needs to be aware of this decision. Check that
_list_of preserves prefixes, for example.

Next up is to continue getting 61, and all supposed-to-be-working
samples, working again.