Add sample 65; needs prefix-of-lists fix.
I propose adding a prefix() function which can handle lists as input.
This makes me wonder if we ever assign a prefix to a non-leaf. I'd like
to avoid that. Next up: fix this test, and create a unit-test todo list
that includes verifying that prefixes are only ever applied to leaves.